+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 0.0.1
 
-* TODO: Describe initial release.
+## 1.1.0
+June 24, 2025
+- T? stateAs<T extends State>() { 
+- static bool setStateOf(StatefulWidget widget, VoidCallback fn) { 
+- static bool rebuildState(StatefulWidget widget) => setStateOf(widget, () {});
+
+## 1.0.0
+July 08, 2025
+- Initial commit